Move CPU tests to cpu_test.rs, prettify CPU code
It's been a chore to constantly jump 400+ lines just to check something I wrote
in the implementation. Sure, this is simplified by vim marks, but extremely
disorienting. It seems to be a common trend for people to be thwarted by tests
that are as long (if not longer than production code), therefore, I'm now
shoving all of it in their own files.

Using google coding styles, all test files will be denoted by *_test.rs.

CPU
 * Generally prettified doc comments.
 * Removed test mod, no need to restate above.