Implement test functions for INC, INX, INY
Finally back!

What seems relatively obvious to me now, but not earlier was that the
INcrement and DEcrement functions dont simply just explode when going
over/under the 8-bit values.

AKA, they have wraparound (my term) arith.
IF 255 is exceeded with IN-/DE-, simply logically AND the value, and
continue on.
 - Tested the idea on Easy 6502 by skilldrick, which concurs.