Implement TAX, TAY, TXA, TXS, TSX, TYA; Increment version to 0.1.3
The transfer functions were as simple as expected, mostly copy
paste. There was the issue of mutable borrow, so to save lines, I made
the set_zn() call rely on the the value being transferred, not the
final value. These should be equivalent values though, so it shouldnt
matter.

Version number is now 0.1.3, next are some general opcodes. These
should actually be somewhat difficult to write, especially the
branching ones.

Added the function skeletons for these v0.1.4 opcodes.