Impl basic parse_opcode function in CPU
Ideally, I would have preferred to have either a function table
lookup, or some other way to lookup function names with the
Instructions struct, but rust doesn't seem to allow this very easily,
for certain varous reasons.

Just making a 255 case long match for the opcodes. It's messy, but
is the fastest option afaik.

So far, just basic copy pasta of opcode and addressing.